+++ include/utils.h
@@ -7,7 +7,7 @@
 
 void config_output();
 void timeout(int miliseconds);
-void get_input(char *input, int size);
+void get_input(char *input, size_t size);
 bool is_blank(const char *input);
 char *to_lower(const char *str);
 void clear_input_buffer();
